@@ -1,8 +1,5 @@
 <?php
-/**
- * @file
- * Load bundle configuration and merge with main config file.
- */
+
 namespace Itk\ExchangeBundle\DependencyInjection;
 
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
@@ -13,12 +10,15 @@ use Symfony\Component\Config\Definition\ConfigurationInterface;
  *
  * @package Itk\ExchangeBundle\DependencyInjection
  */
-class Configuration implements ConfigurationInterface {
-  /**
-   * {@inheritDoc}
-   */
-  public function getConfigTreeBuilder() {
-    $treeBuilder = new TreeBuilder();
-    return $treeBuilder;
-  }
+class Configuration implements ConfigurationInterface
+{
+    /**
+     * {@inheritDoc}
+     */
+    public function getConfigTreeBuilder()
+    {
+        $treeBuilder = new TreeBuilder();
+
+        return $treeBuilder;
+    }
 }
