@@ -9,8 +9,8 @@ namespace Itk\ExchangeBundle\Exceptions;
 /**
  * Class ExchangeSoapException
  *
- *
  * @package Itk\ExchangeBundle\Exceptions
  */
-class ExchangeSoapException extends \RuntimeException {
+class ExchangeSoapException extends \RuntimeException
+{
 }
