@@ -7,62 +7,68 @@
 namespace Koba\AdminBundle\Controller;
 
 use FOS\RestBundle\Controller\FOSRestController;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Symfony\Component\Routing\Annotation\Route;
 use FOS\RestBundle\Controller\Annotations as FOSRest;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use FOS\RestBundle\Controller\Annotations\View;
-use Symfony\Component\HttpKernel\Exception\HttpException;
 
 /**
  * @Route("/resources")
  */
-class ResourceController extends FOSRestController {
-  /**
-   * Get resources.
-   *
-   * @FOSRest\Get("")
-   *
-   * @View(serializerGroups={"admin"})
-   * @return array
-   */
-  public function getResourcesAction() {
-    return $this->get('itk.exchange_service')->getResources();
-  }
+class ResourceController extends FOSRestController
+{
+    /**
+     * Get resources.
+     *
+     * @FOSRest\Get("")
+     *
+     * @View(serializerGroups={"admin"})
+     * @return array
+     */
+    public function getResourcesAction()
+    {
+        return $this->get('itk.exchange_service')->getResources();
+    }
 
-  /**
-   * Get resource by mail.
-   *
-   * @FOSRest\Get("/{mail}")
-   *
-   * @View(serializerGroups={"admin"})
-   * @return Resource
-   */
-  public function getResourceByMailAction($mail) {
-    return $this->get('itk.exchange_service')->getResourceByMail($mail);
-  }
+    /**
+     * Get resource by mail.
+     *
+     * @FOSRest\Get("/{mail}")
+     *
+     * @View(serializerGroups={"admin"})
+     * @return Resource
+     */
+    public function getResourceByMailAction($mail)
+    {
+        return $this->get('itk.exchange_service')->getResourceByMail($mail);
+    }
 
 
-  /**
-   * Refresh resources.
-   *
-   * @FOSRest\Post("/refresh")
-   */
-  public function refreshResources() {
-    $this->get('itk.exchange_service')->refreshResources();
-  }
+    /**
+     * Refresh resources.
+     *
+     * @FOSRest\Post("/refresh")
+     */
+    public function refreshResources()
+    {
+        $this->get('itk.exchange_service')->refreshResources();
+    }
 
-  /**
-   * Update resource alias
-   *
-   * @FOSRest\Put("/{resourceMail}/alias")
-   *
-   * @param Request $request
-   * @param string $resourceMail
-   */
-  public function setResourceAlias(Request $request, $resourceMail) {
-    $resource = json_decode($request->getContent());
+    /**
+     * Update resource alias
+     *
+     * @FOSRest\Put("/{resourceMail}/alias")
+     *
+     * @param Request $request
+     * @param string $resourceMail
+     */
+    public function setResourceAlias(Request $request, $resourceMail)
+    {
+        $resource = json_decode($request->getContent());
 
-    $this->get('itk.exchange_service')->setResourceAlias($resourceMail, $resource->alias);
-  }
+        $this->get('itk.exchange_service')->setResourceAlias(
+            $resourceMail,
+            $resource->alias
+        );
+    }
 }
