@@ -12,8 +12,10 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
  * Class KobaAdminBundle
  * @package Koba\AdminBundle
  */
-class KobaAdminBundle extends Bundle {
-  public function getParent() {
-    return 'FOSUserBundle';
-  }
+class KobaAdminBundle extends Bundle
+{
+    public function getParent()
+    {
+        return 'FOSUserBundle';
+    }
 }
