@@ -3,6 +3,7 @@
  * @file
  * Contains the command for ConfirmBookingCommand.
  */
+
 namespace Koba\MainBundle\Command;
 
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
@@ -16,44 +17,50 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
  *
  * @package Koba\MainBundle\Command
  */
-class DeleteBookingCommand extends ContainerAwareCommand {
-  /**
-   * Configure the command
-   */
-  protected function configure() {
-    $this->setName('koba:booking:delete')
-      ->addArgument(
-        'id',
-        InputArgument::REQUIRED,
-        'Which booking entity to delete?'
-      )
-      ->setDescription('Remove booking from Exchange');
-  }
-
-  /**
-   * Executes the command
-   *
-   * @param InputInterface $input
-   * @param OutputInterface $output
-   *
-   * @throw NotFoundHttpException
-   *
-   * @return void
-   */
-  protected function execute(InputInterface $input, OutputInterface $output) {
-    $container = $this->getContainer();
-    $doctrine = $container->get('doctrine');
-
-    // Get booking.
-    $id = $input->getArgument('id');
-    $booking = $doctrine->getRepository('ItkExchangeBundle:Booking')->findOneBy(array('id' => $id));
-    if (!$booking) {
-      throw new NotFoundHttpException('booking with id:' . $id . ' not found');
+class DeleteBookingCommand extends ContainerAwareCommand
+{
+    /**
+     * Configure the command
+     */
+    protected function configure()
+    {
+        $this->setName('koba:booking:delete')
+            ->addArgument(
+                'id',
+                InputArgument::REQUIRED,
+                'Which booking entity to delete?'
+            )
+            ->setDescription('Remove booking from Exchange');
     }
 
-    // Check Exchange to see if the booking has been accepted.
-    $exchangeService = $container->get('itk.exchange_service');
+    /**
+     * Executes the command
+     *
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     *
+     * @throw NotFoundHttpException
+     *
+     * @return void
+     */
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+        $container = $this->getContainer();
+        $doctrine = $container->get('doctrine');
 
-    $exchangeService->cancelBooking($booking);
-  }
+        // Get booking.
+        $id = $input->getArgument('id');
+        $booking = $doctrine->getRepository('ItkExchangeBundle:Booking')
+            ->findOneBy(array('id' => $id));
+        if (!$booking) {
+            throw new NotFoundHttpException(
+                'booking with id:'.$id.' not found'
+            );
+        }
+
+        // Check Exchange to see if the booking has been accepted.
+        $exchangeService = $container->get('itk.exchange_service');
+
+        $exchangeService->cancelBooking($booking);
+    }
 }
