@@ -3,6 +3,7 @@
  * @file
  * Contains the command for updating the Calendar events.
  */
+
 namespace Koba\MainBundle\Command;
 
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
@@ -14,28 +15,31 @@ use Symfony\Component\Console\Output\OutputInterface;
  *
  * @package Koba\MainBundle\Command
  */
-class UpdateCalendarCommand extends ContainerAwareCommand {
-  /**
-   * Configure the command
-   */
-  protected function configure() {
-    $this
-      ->setName('koba:calendar:update')
-      ->setDescription('Update calendar information from XML file.');
-  }
+class UpdateCalendarCommand extends ContainerAwareCommand
+{
+    /**
+     * Configure the command
+     */
+    protected function configure()
+    {
+        $this
+            ->setName('koba:calendar:update')
+            ->setDescription('Update calendar information from XML file.');
+    }
 
-  /**
-   * Executes the command
-   *
-   * @param InputInterface $input
-   * @param OutputInterface $output
-   *
-   * @return void
-   */
-  protected function execute(InputInterface $input, OutputInterface $output) {
-    $calendarService = $this->getContainer()->get('koba.calendar_service');
-    $calendarService->updateXmlData();
+    /**
+     * Executes the command
+     *
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     *
+     * @return void
+     */
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+        $calendarService = $this->getContainer()->get('koba.calendar_service');
+        $calendarService->updateXmlData();
 
-    $output->writeln('Calendar events updated.');
-  }
+        $output->writeln('Calendar events updated.');
+    }
 }
