@@ -7,21 +7,23 @@
 namespace Koba\MainBundle\Controller;
 
 use FOS\RestBundle\Controller\FOSRestController;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Symfony\Component\Routing\Annotation\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use FOS\RestBundle\Controller\Annotations as FOSRest;
 
 /**
  * @Route("")
  */
-class IndexController extends FOSRestController {
-  /**
-   * Get index page.
-   *
-   * @FOSRest\Get("")
-   * @Template()
-   */
-  public function indexAction() {
-    return array();
-  }
+class IndexController extends FOSRestController
+{
+    /**
+     * Get index page.
+     *
+     * @FOSRest\Get("")
+     * @Template()
+     */
+    public function indexAction()
+    {
+        return array();
+    }
 }
